auth_service/pyproject.toml
[tool.poetry]
name = "expo-auth-service"
version = "0.1.0"
description = ""
authors = ["bebra_dev"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.112.0"
sqlalchemy = "^2.0.31"
psycopg2-binary = "^2.9.9"
pydantic-settings = "^2.4.0"
alembic = "^1.13.2"
asyncpg = "^0.29.0"
pyjwt = "^2.9.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
uvicorn = "^0.30.5"
pydantic = {extras = ["email"], version = "^2.8.2"}

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
black = "^24.8.0"
pytest = "^8.3.2"
testcontainers = {extras = ["postgres"], version = "^4.7.2"}
httpx = "^0.27.0"
pytest-asyncio = "^0.23.8"
pytest-mock = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length=79
skip-string-normalization=1
extend-exclude='''
(
    \.git
  | \.mypy_cache
  | \.venv
  | migrations
)
'''

[tool.isort]
default_section = "THIRDPARTY" 
known_local_folder = ["db", "users"]
sections = ["FUTURE","STDLIB","THIRDPARTY","FIRSTPARTY","LOCALFOLDER"] 
skip = [".gitignore", "venv", "env"] 
skip_glob = ["*/migrations/*"] 
multi_line_output = 3
line_length = 79
lines_after_imports = 2
include_trailing_comma = true